feat(planner): integrate calendar with real data and update header reactivity

+++ lib/features/planner/providers/stats_provider.dart
@@ -19,7 +19,7 @@ final monthlyTaskStatsProvider = Provider<MonthlyStats>((ref) {
       return t.date!.year == monthDate.year && t.date!.month == monthDate.month;
     }).length;
 
-    if (monthTasks > maxCount) maxCount = monthTasks;
+    if (monthTasks > maxCount) maxCount = monthTasks.toInt();
     totalCount += monthTasks;
 
     // Format label (e.g., JAN, FEV)
